expunges: add v4.19.17 and v4.19.18 baselines for xfs
This adds a baseline for xfs for the stable kernels v4.19.17 and
v4.19.18. This effort is to help address the set of stable fixes
which will be sent out.

The v4.19.18 baseline is copied from the v4.19.17 baseline, there
are no changes, I haven't yet completed the v4.19.18 baseline but
I expect it will be the same. This will be confirmed later and
modified / expanded if new failures are found. Given that no changes
have gone in for stable for v4.19.y we expect the baseline to not have
changed yet.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>