qemu-kernel-confis: add CONFIG_I6300ESB_WDT=m
Enable the recommended emulated qemu watchdog on the kernel.
We use this to be able to trigger a reset of the device in
case of failure. On your libvirt guest config, you will
need the following entry added to each guest:

sudo virsh edit xfs

And then add, right after the <devices> tag:

<watchdog model='i6300esb' action='reset'>
<address type='pci' domain='0x0000' bus='0x00' slot='0x0b' function='0x0'/>
</watchdog>

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>