oscheck: add initial ansible and vagrant support
This adds ansible support to use oscheck, and to let us install a custom
kernel easily. This lets us parallelize testing fstests per target
filesystem configuration we wish to support.

The however still assumes you build your own virtual servers...
This is addressed with vagrant support. It lets us provide the
recipe for virtualization needs on a configuration file. We only
support libvirt and virtualbox for now. OpenStack and Azure
should be easily possible to add with terraform scripts.

For now this just lets you demo the setup with vagrant:

cd vagrant
vagrant up

That will provision with ansible only developer configuration
work for now. Later on we'll move on to test using the ansible
playbook for oscheck and bootlinux.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>