oscheck: add initial basic terraform support with azure
This adds a basic initial terraform setup with Azure.
We can expand on this later, but it shows the potential.

With dynamic configuraitons for hosts this will be much
easier. Unfortunately at least for the free trial on Azure
it takes about over 1 minute to provision just one guest,
and about 2.5 minutes to destroy it. Not ideal IMHO, but
I haven't tested other cloud systems yet.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>