terraform: add multiple host support for azure provider
This expands on the azure provider to add the promised
support for multiple nodes. Note that we force the limit
to 4 nodes for now, only because if you are like me and just
trying the demo Azure subscription you'll end up with a limit
of only creating 4 cores / nodes. Otherwise you'll run into
a resource limit error while creating the fs tests nodes.

This depends on the terraform yaml plugin, we use this to
easily let us have a single point of entry for the nodes
we want to use. This lets us use for instance the xfs.yaml
file for both vagrant and terraform.

I was a bit lazy and just used count.index for appending
the names of resources we didn't care for too much to ensure
they are in sync with the hostnames, but this can easily
be adjusted. Its just that the way we currently capture
the hostname is a bit hacky. It would be cleaner to leave
the hacks to a local variable somehow and then extract the
hostname in a clean way through a clean variable. This way
if the *hack* for the names is adjusted later, we just have
adjust the hack *one* insteade of multiple times.

This also fixes printing the IP address of the target nodes,
but we do so by printing all the IP addresses. We can later
put the hostname next to the IP but this should do for now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>