vagrant: name nvme disks something more useful
Provide names to the nvme disks we strive to create so
that its a bit clearer the intention behind them. This
is actually also requiered to prevent the terraform
yaml plugin from crashing on us, it has a bug where if
the name is a digit it crashes.

This requires a small change on the Vagrantfile to not
rely on the key for the port number. This is rather easy.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>