terraform/aws: add elastic block storage support
The free trial restricts us to only the t2.micro instance type,
which only has 8 GiB of storage. We add Elastic Block Storage (EBS)
support so we can request what we need on the fly.

The free trial restricts us to.. 30 GiB... However I just noticed
I was allowed to create up to 150 GiB on one try... wonder if I got
charged or if the free trial got a bump on storage.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>