vagrant: fix provider loop by using colon to avoid multiple loops
If you use a provider and then embed another loop underneath it
x times, and then another y times it would seem you end up looping over
the provider x * y * y  times instead of x * y. In our case
we see that we loop over the provider on the *same* host for each
disk, the number of disk there are present. In practice this ends
up with an error after we've created hosts, shut them down and then
try to bring them up again via:

   vagrant up

We get an error from storagectl call for virtualbox telling us that
the storage controller on port 0 is already present.

It took a good while to realize that the issue was the missing
colon on the provider loop. Without this we go bananas.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>