oscheck: mkfs the TEST_DEV partition if ./gendisks.sh -m is used
It can easily be forgotten to mkfs the TEST_DEV, so let gendisks.sh
mkfs it for us if its called with -m.

The idea is to later use -m via ansible for the setup, so that we can
easily create the setup needed prior to extending the oscheck playbook
to start running tests easily.

This also allows us to easily add other filesystems by setting the
FSTYP argument 'make install' target time. Even though the top level
configuration file lets us support different filesystems, for now
we stuff our favorite XFS configuration file to its own file. We can
evaluate later if we want to keep these separate or not.

Lastly, since we are now providing *new* TEST_DEV requirements for
the new -m option just replace the configuration file *and* the
.oscheck-setup file if install target is used.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>