kdevops: fork oscheck and then trim and rebrand to kdevops
It turns out that the infrastructure being developed for
virtualization/cloud can be re-used / re-purposed by many
projects so just fork the original oscheck code and trim
it down with a new purpose.

Later on oscheck then can be come a *user* of this infrastructure.

We start out with *one* use case:

  * devops for kernel developers: first test case is oscheck for
    fstests -- Linux filesystem testing.

We can add other uses case demos later, but for now this is our
first target.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>