bootlinux: modularize partition creation
I was observing that /data/ was sometimes not mounted after
running the bootlinux role. I then realized its likely /data/
as a partition was *not* being used on a fresh system as no one
was creating it. We only created it once we got up to the
oscheck role.

The oscheck role creates two partitions:

  * /data/
  * /media/truncated

The boolinux role needs:

  * /data

As much code should be shared as is possible. This means we need
to generalize the creation of paritions, and also then generalize
the creation of /data/ and let both oscheck and bootlinux roles
use it.

We create a role for partition creation, and one for creating
/data/ and then make oscheck and bootlinux role use the /data/
partition creation one.

We're careful to not create / mkfs the partition unless its not
present yet.

For this commit only address bootlinux role and the creation of
the partitions generic work, a separate commit will be for oscheck.
This is done so that this commit can be integrated as-is into
kdevops.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>