create_partition: provide paranoia partition guards
One cannot assume that the same /dev/nvme0n1 will be the same
/dev/nvme0n1 on the second boot. The current create_partition
role is too loose, and will always create a target labeled
partition if one does not exists on the specified device.

This means you end up duplicating data on different drives,
and depending on your luck, you may or may not get your data
upon reboot. If you don't you end up re-creating the whole
tamale.

Fix by providing three guards and *iff* all three are false can
we then mkfs your partition for you. The guards which *all* must
be met are:

  a) target mount point is not mounted
  b) target mount point is not present on /etc/fstab
  c) target filesystem label does not exist on the system

If all you were lacking was the a) and b) but had c), that is
the filesystem was created, existed with the respective label,
but had no /etc/fstab entry we'll create it for you and mount it
if and only if of course it matches the same original target device
name. If it doesn't update your specification.

The point to this is you *do* create the target partition only
once and with the device target in mind. It may change device name
later but the label will still work and mount the device.

The last check c) will ensure that we don't re-create a filesystem
target if one was already present with the system with the same
label.

If you're manually screwing around with things manually you're
on your own, but these safeguards should suffice for most dynamic
device changes.

To test things you may want to add a tag for you respective
partition, and then also use the 'partition' tag in this role.
For instance:

ansible-playbook -i hosts -l dev -l oscheck-xfs-dev --tags partition,data_partition bootlinux.yml

This should be a no-op after the partition is created. And a no-op after
all further reboots, unless you kill the label, (wipefs -a), the
/etc/fstab entry, and umount the target path.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>