vagrant: use libvirt-qemu group as default and allow to override
When testing kdevops on Debian testing I ran into a series of permission
issues. The first issue was that the group used is libvirt-qemu. The
second was apparmor.

Set the default group to libvirt-qemu as we support debian out of the
box first. Allow to override the group settings without having to
require modifying nodes.yaml by using the environment variable:

  * KDEVOPS_VAGRANT_QEMU_GROUP

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>