vagrant: add safety check for limit number of boxes
If you specified in your environment variable a higher number
of boxes than what we define you would never end up running ansible
playbooks, as we only run them when we know we're on the last one.

Fix this by setting the limit based on the min between the max allowed
and what the user specified.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>