vagrant: add project namespace inference support
Syncing between Vagrant files updates was difficult due to clashes
with project names. Instead of dealing with that nonesense, assume
your project name is based on the parent directory where the vagrant
directory exists.

So now... you can sync with the latest Vagrantfile on external projects
without having to deal with conflicts.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>