bootlinux: publish role to ansible-galaxy and use it
This means we also can remove now the create_data_partition role
as bootlinux now open codes it.

This also bumps the create_partition to v1.2.0 given that prior to this
release by default the mount options were left empty, and when that is
done the /etc/fstab gets populated incorrectly. With the v1.2.0 the
default mount option is "defaults" as is expected.

This goes tested with a full fetch of linux, compile, install, reboot
and confirmation we are on the freshly compiled v4.19.62 kernel.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>