terraform: do not rely on vagrant for the terraform node configuration
Do not rely on users having a vagrant setup for the terraform stuff.
So just require the user to have the file present. This will make it
easier once we make this an ansible role.

I can't see why we need the project namespace stuff for the prefix for
the node configuration file as we did for vagrant, so just don't require
it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>