kdevops: fix dhcp races when using linux-next
When booting the latest linux-next I see that network interfaces are up
but dhcp never kicks in with Debian. It seems we have yet again a device
name change issue... Seriously...

If we are not doing network development or working on systemd stuff we
frankly don't care about this bug, so just allow us to work around it.
*We* kdevops kernel developers don't care about this bug so just
work around this nasty issue by ensuring all of our non loopback devices
*get* a DHCP request issued, *always*.

We use a new ansible role for this which captures this easily.
Only observed on Debian so far, so other distros won't have anything
done for them.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>