Makfile: use --force for installing requirements
It seems that the feature I want, to update roles which you have a need
to update *won't* be implemented by design on ansible. I read the
feature request on github and it seems this went no where. I have no
hope in that moving forward so just adapt to the scheme which ansible
like to promote. Force you to download all roles.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>