Simplify by using kdevops_install role
We typically recommend forking kdevops, but that means that
it is not easy to upgrade. Likewise, uprading typically means
copying the requirements file, and the set of playbooks which
we may have adopted. This doesn't scale.

Note, I am aware of galaxy collections, and using that would
be indeed be nicer, however that requires setting up a token,
and then using a "Red Hat Automation Hub" mombo jumbo. No
thank you.

What we *loose* by not using Galaxy collections is an easy way
to just let us gather a set of roles of a version. We instead
have to *force* install all needed roles whenever someone
will call 'make deps'. I can live with that compromise and
blame ansible's lack of flexibility requiring the other fluff
fluff.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>