Update kdevops to v1.9.8
This update kdevops to v1.9.8 which provides a few role updates
to ensure we get debian testing working just fine. The updates
includes the ability to specify the ansible python interpreter,
and we update the devconfig role to remove the xfsdump package
which is now replaced by just using xfsprogs on Debian.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>