vagrant/kdevops_nodes.yaml: remove ssh key reference
We don't do anything with this, so just remove it. We also
don't encourage static keys, there is no need for them. It is
a security risk if you use static keys for a set of hosts, and
then that key gets compromised. We support by default creating
random keys and then updating your ~/.ssh/config for you, so
it can't get any easier than this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>