Makefile: fix use of the force configuration option
Without this we were never updating the ansible roles.
So for instance, even though you may have gotten the
latest kdevops_install role, you would not have gotten
the latest bootlinux role. Right now the bootlinux
role is at v2.0.1 and the v2.0.0 release was using
the v4.19.62 kernel, which had a compilation issue.

The v2.0.1 release bumps this up to v4.19.133.

Fixing the force will ensure we *upgrade* as well.
I really really really *do not like* the ansible
"feature" by which --force is required to do *updates*.

The reluctance by ansible upstream to not pay attention
to the critique that this is needed, is simply wrong.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>