kconfig: add support for options for configuring bootlinux
Instead of carrying around custom variables in extra files, we can
just add support for the different configurations we already support
and use.

It should be fairly simple for folks to extend the default then with
their own things with this. Likewise, this should also let you easily
extend this to add custom URLs if you fork this project.

We make linux-next the default as that is where I do development.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>