README.md: reorder the file a bit
Now that kconfig is supported, my *own* experience is that
things are *incredibly* simplified. So most of the details of
how to run ansible, etc, is just fluff that user don't need
to care about.

Sure, developers may care... but, even us, we won't care after
we get an easy way to tweak knobs with kconfig.

Experience is also showing that we can easily in the future just
have targets to run tests. Example future targets I can think
that we might add:

  * make seltests
  * make fstests
  * make blktests

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>