Update to kdevops v3.1.0
Terraform has been updated so to share the ssh configuration variable
names of both username and the ssh pubkey file we'll use.

One of the things we should strive for next, is to createa a new
random ssh key for the user if the one specified by does not exist.

The variable name fo the ssh key file is: ssh_config_pubkey_file

And the default is ~/.ssh/kdevops_terraform.pub.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>