README.md: split documentation into sections
We had tons of documentation about *how* to use kdevops manually.
With kconfig, manual operation is simply overkill, it has allowed
us to hide complexity for the user through Makefile targets, and
by creating configuration files for use.

The documentationa about *how* all this works however should not be
lost, so add a new section which expands on this as if this were
documentation intended for understanding under the hood of kdevops.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>