README.md: clarify rolling Linux distribution recommendation
The complexities which bring about encouraging a rollinux Linux
distribution frankly comes about because of the use of the
vagrant-libvirt plugin, and dependency hell this can bring upon.

Experience shows, for instance, that trying to use vagrant-libvirt
on at least SUSE Enterprise releases ends up in a very bad experience.
Resolving this is not easy, and frakly not worh it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>