Makefile: fix chicken-and-egg problem with Makefile.kdevops
The Makefile.kdevops file is deployed for us as part of the
build process, however we are currently chaining in dependencies
after this, and they are not yet defined. We currently tried to
optionally include Makefile.kdevops, that is, if it does not
exist, obviously these targets which we rely on won't exist.

And so currently we have to run make twice...

Fix this by creating two stages to the build process.

Using the whole obj-y stuff was cute, so folks understand how that
works, but its not allowins us to make things clear. So provide
more sensible target names for these targets.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>