kconfig: add boot linux data partition specification support
This adds the ability to be able to specify *how* to create the
partition used to for to place the Linux git clone we are doing.

Right now this works well only on vagrant as the defaults with
vagrant work. But for terraform we have to give much more flexibility.
For instance with AWS the default is to boot into a PV guest and
the device exposed is througha EBS, and although this can vary
depending on the *type* of guest sicne we're only supporting
one right now, and we only have cloud support through kconfig
for *one* cloud yet, we default to /dev/xvdh for AWS.

Future cloud support should be easy to add, once we figure out
which partitions are created. Vagrant defaults to using an nvme
drive, we do this as well, the nvme code is much nicer than
anything in existence before. scsi... oh my... Anyway, even
though nvme should technically be trivially be cheaper for any
cloud solution, cloud solutions tend to "sell" the idea you
are using real nvme when using nvme on guests. And so we can't
always rely on the default to use nvme.

So allow the user also to specify what they want to use.

Note: we don't currently allow disabling this feature of
creating this data partition, but it should be easy to
add support for it. However if you are are compiling linux
for development purposes even on the cloud you likely
want to keep with this tradition, as it lets you pick much
more expensive drives for *fast* IO. That's all we care about.

Hell, if you can afford it, go for RAM devices!

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>