azure: add new vm image Standard_DS3_v2 and make it the default
The Standard_DS1_v2 is too puny, so use Standard_DS3_v2 as the
default. Also, I should note that even if you pay Azure the
default max number of cores you can spawn is 10, and so if
you are spawning two guests we are still safe with this
option as we'd have 4 vpus per guest, so 8 cores.

Sadly, the Azure forces you to *spawn* a ticket if you want
to spawn *more* vpus and you have to *actually* interact
with a human being over email explaining you are using
terraform and that you want to increase your quota over 10
cores....

Insanity!

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>