kconfig: enable post bring up ansible provisioning to be configurable
We didn't enable ansible on terraform solutions yet as we had not
figured out how to ensure all cloud providers had the the user's
ssh configuration file updated (~/.ssh/config), but now that I've
found a workaround with Azure, it puts us on a path to enable
ansible all around.

Not all users may want to enable ansible to be used post bring up
though, and so enable users to disable this.

Looking over the menu though, there are two ansible sections, and so
two ways in which we use ansible:

 * installing kdevops's kdevops_install ansible role, and then using
   that to bring up and install all the other kdevops ansible roles

 * use ansible to provisiong the hosts created at 'make bringup'

So, to avoid confusion from users split the menu up so that this
is clear.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>