bootlinux: add support to use user and group inference
Providing sensible defaults for the target user/group of the
partition we are going to use to build Linux is pretty cumbersome.
Instead of extending our list of known defaults every time we add
a new image to support, add instead a feature which lets the
bootlinux role infer the target user/group based on the default
logging settings.

We enable this by default.

Then, if users want to really customize this to some other settings,
they just turn on manual drive.

To take advantage of this feature we bump the kdevops version, so
we can get the new bootlinux role.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>