requirements: update kdevops to v3.2.0
This fixes a devconfig role issue where I noted that ssh stopped
working after using the provisioning role on either azure or gce.
The issue, it turns out, was that we update sources and install
a few packages only. This seems to have left out upgradeing ssh
and very likely ssh had a broken library, which the dependency
mapping didn't pick up. Doing an upgrade manually on pre-established
terminal after the issue comes up fixes the issue.

And so to avoid these issues from creeping up in the future the
devconfig role now does a full upgrade of all packages found
if we are doing a refresh on the list of packages available
or installing new packages after doing a source frefresh.

The terraform role also got updated, to enable gce to jump on
the bandwagon of enabling ansible provisioning. Now all we
have to do is to do the kconfig magic for gce.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>