Makefile: add support to build extra_vars.yaml automatically
The ansible kdevops roles have sensible defaults for vagrant,
but for terraform we end up having to modify things quite a bit.
One is the target drive, another is the target user, and also the
group owner for files.

The target user / group owner for files is used by the bootlinux
role, but I've extended it to infer your user/group using getent.
Similar thing can be done with devconfig. But I'm lazy right now,
and this at least demo's *how* we can build the extra_args.yaml
file for users.

Note that you typically have to call ansible with the following
argument if one wanted to just run the devconfig playbook on
just the kdevops host:

  ansible-playbook -i hosts -l kdevops --extra_args=@extra_args.yaml

All kdevops files look for the extra_args.yaml file automatically,
and so this is not needed. So all you'd need to do if provisioning
failed on terraform:

  ansible-playbook -i hosts -l kdevops

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>