kconfigs: add fstests as a new kdevops workflow
I've gone back and forth now many times between whether or not
we should follow similar logic on *forking* kdevops for each
workflow as we used to recommend before. But with kconfig that
seemed just pointless now. But the integration of what we was
'oscheck' into kdevops seemed difficult as well... Likewise,
updating 'oscheck' to use the latest 'kdevops' with kconfig
was a lot of work as well...

It seriously took me about 1 or 2 weeks to decide *how* to try
to integrate what was 'oscheck' into kdevops, or bring 'kdevops'
to the 'oscheck' tree.

What ultimately was saner was to try to merge oscheck into kdevops,
since that would allow easier visibility of what was happening on our
ledger in terms of *evolging* kdevops which is more of what we cared
for, and it also would demo how to add a very big workflow into our
tree. The amount of work just seemed daunting and I could not grasp how
to do this right.

This re-implements what was 'oscheck' as a completely new workflow on
kdevops.  This means we get to keep the lineage of work of where kdevops
came from, and we simply merge back to this tree the evolution of trying
the concept of 'kdevops' in 'oscheck'.

This should hopefully make it clear to others what is expected and
*how* to integrate a complex test environment into kdevops as a
new workflow.

At this point, I can only verify that this works to the point I can
get debian to start running oscheck.sh. The baseline is definitely
outdated, which means my guests will likely crash... unless of course
the debian testing kernal has 0 flaws and the general expunge list
didn't skip any known issues.

We'll now find out, and with this will follow a set of example patches
of *how* to extend the baseline for debian testing's kernel. This is
only *begging* for automation of a new task:

  * give me a baseline for a new random kernel $BLEH

With a whole lot of watchdog love, this should in theory be possible.

Worth noting also, I am only testing debian testing using the default
section. Nothing more.

A lot of the 'oscheck' prefix names for variables were traded in for
an 'fstests' prefix, where it was deemed that these things likely belong
inside fstests in the future.

The questions then now become:

  - What from kdevops should we merge back to fstests?
  - How do we slim kdevops further? What implications does this have?
    Example of things to consider:
      a) kconfig
      b) baselines and who provides them
  - Package name difference mess, who's responsibility / goal is it to
    unify this question, in way in which developer no longer need to
    care?
  - How do we establish a new baseline for a known filesytem *faster*?
  - How do we establish a new baseline for an unknown filesytem *faster*?

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>