fstests: update and fix support for processing results
The oscheck approach had all "fstests" files locally, however
when merging "oscheck" into kdevops as a workflow we needed
to change things a bit around. So adjust the scripts / paths
used.

This also adds support for Python3 and makes it explicit because
that is what most modern distributions are using now.

Result processing is done on the command and control node, and so
we need a way to install dependencies for the tools we use to
process results. To do this we could either squeeze the tasks into
the fstests role, but that makes things odd given that the distro
check is for the target node, no the local controlling node. The
semantics would get really odd if we used local_action with and need
to extend this later. In order to avoid sloppy semantics we add
a dedicated role for the local deps on the controlling node.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>