fstests: add ability to do manual section for XFS
When you create an XFS filesystem you have certain
features / options which you can enable / disable.
A lot of this is features which get added over time
and as they stablize are eventually enabled by default
in the generic 'mkfs.xfs' command. As features get
enabled by default, this means newer distributions
must use configuration options manually to replicate
or createa a filesystem compatible with older kernels
which may not have certain features. There are also
some options which may never get enabled by default,
an example was the realtime device.

In order to test the full array of all features that
we have for XFS we currently test 7 different
configurations. We use an fstests section to represent
each section, and kdevops creates a node per section.
That node will run tests for is own section.

Additionally, we may want to have a setup where a
kernel developer will have a host available to test
a known baseline, but also a new kernel to ensure
there are no regressions, both live at the same time.
For this reason we support creating by default both
a node used for the baseline and another for
development purposes.

The full setup of testing requires 14 nodes. Even if
a Linux distribution wanted to do its best and test all
valid popular sections, some distributions disable certain
features like support for the realtime device. Likewise,
we want to enable kernel developer to ramp up and test
only a target desirable section. Not all kernel developers
will have access to a system to create 14 guests to run
all tests, for example.

This adds support to do manual selection of the individual
XFS sections you may need to run tests for through kconfig.
It also allows developers to enable / disable adding an
extra node for development purposes. This would be disabled
by QA-only type of systems where you are sure developers are
not expected to log in, and try to debug without access to
a both a baseline and a development system.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>