fstests: move kconfig to workflows/fstests/Kconfig
For ansible workflows which we are not downloading via
ansible galaxy but host oursevles in-house, use the
direct workflows/fstests/Kconfig file instead of
the abstract kconfigs/workflows/fstests/Kconfig.

We'll keep on using kconfigs/workflows/* path for
ansible workflows for which we *do* download via
ansible galaxy though.

This is a little more intuitive, as folks looking
underneath the workflows/fstests/ directory can
find most things associated with it there. The
only thing left outside of that, is the scripts
and python directory, but it makes sense to keep
a different directory for those.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>