fstests: provide groupings for hosts on kconfig
There are a couple of non-crc enabled filesytems we can
run tests on, likewise we have a couple of reflinked enabled
fileseystem we can run tests on. Provide a group for these two.

This will make it easier for the folks configuring tests to
figure out how / why things should be enabled. We'll expand
shortly on the reflink enabled filesystem too, so this will
allow better grouping / understanding of where these target
filesystem belong.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>