xfs: fix disabling big block section
When manual selection is not enabled we currently always include
the big block section. This was party due to a mistake on the
script which was checking for the wrong thing. However, if we just
check for the configuration option for it, it would not exist on
ppc64le or aarch64 unless the manual selection has been enabled.
To address this we must create a hidden kconfig entry for when we
*don't* enable manual selection and always default it to y.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>