kdevops: add support for custom host prefix
If you use kdevops twice on a system you will run into a
conflict. We need a way to let the user customize the hostnames
used. This applies to baremetal, local vms (vagrant) or cloud
solution (terraform).

Likewise, some distributions may want to use a prefix to hint
what release you are working with. This will allow distributions
to add their preferred prefix for select images.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>