vagrant/Kconfig: display the vagrant box version only when needed
A few vagrant box images do not need a box version specified,
but displaying the option to those users may confuse them.
Only enable the version option when a vagrant box image requires
a version specified. If we know we don't need it, don't bother
displaying the option on the menu.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>