workflows/fstests/xfs/Kconfig: provide HAVE helpers enabling sensible defaults
Distributions may decide to add support to a feature upstream, or, with
time, they may deprecate features. These decisions are pretty important
for the support chain for a distribution. Once a feature is supported,
it must be tested. Likewise, a deprecated feature means you may no
longer want to bother testing for it.

In XFS we have had over the years new developments for new features,
and likewise, we've also have had odd old features which some
distributions have decided not to support.

To support these decisions and to be able to reflect them easily on
the fstests xfs configuration we introduce a new set of HAVE* kconfig
symbols:

  * HAVE_DISTRO_XFS_PREFERS_*
  * HAVE_DISTRO_XFS_SUPPORTS_*
  * HAVE_DISTRO_XFS_IGNORES_*

These provide enough semantics for kconfig so that we can provide
sensible defaults when selecting a distribution image to test. The
distribution would simply select these HAVE* symbols so that features
are disabled/enabled by default on the XFS configuration upon selection
of these images.

This let's the QA engineer / developer testing not have to think about
what is supported or not when selecting the image. This becomes a coded
information through kconfig. They obviously can override these. A
different symbol is used instead of using the direct per-section XFS
symbol, because if we were to use the symbols directly the user would
not be able to unselect them. With HAVE*( symbols these are *hints*
towards enabling/disabling by default, they are not forced when
selecting an image.

No distribution yet uses these, but soon these will be used.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>