kconfigs/Kconfig.kdevops: add support for a custom host prefix
Often times you are testing a specific distribution release,
and you want to identify the targets as such. Allow for a
distribution to override the defaults "kdevops" prefix, with
whatever it prefers. The way to use this is to set the prefix
with a variable, if the distribution relaese is selected.

Furthermore, not all distributions want to use this prefix by
default, but some do prefer it. To ensure it is not enforced
we add helper HAVE* symbols to let the distribution enable the
default to use the prefix or not. This still allows users to
enable this mechanism even if the distribution doesn't prefer it
and for it to be set.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>