vagrant/Kconfig: add support for private vagrant box URL
Some distributions may have boxes avaialable only on custom
URLS, and have no easy way yet to download them off of the
official vagrant registry. For these we must support the ability
to download boxes using a custom URL. Add support for this.

The vagrant images which require this can select HAVE_VAGRANT_BOX_URL.
This is currently unused but we'll expand on this later.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>