vagrant: add initial boilerplate support for SUSE Linux Enterprise
The vagrant images for SUSE Linux Enterprise are not on the public
vagrant registry, you however can download them via:

https://www.suse.com/download/

But there is no programatic way to download these yet. It is being
evaluated how to make this easier, if possible. In the meantime, this
adds boilerplate needed to easily add SLE support by looking for the
kconfig file for SLE, and if its present and the file is not empty
we process it.

Kconfig doesn't have a way to dynamically source a file, even if you
use a kconfig symbol to selectively source it, kconfig still requires
that file to be present. So our version of the SLE kconfig is empty
for now. This allows the file to be extended as its being evaluated
how this can be supported / if at all publicly.

SLE becomes the first release to also use rely on the private box
installation mechanism.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>