kdevops: enable twoline method distribution registration system
Some distributions as SUSE Linux Enterprise can make use of registation
system in order to make the system more useful. We define a simple
twoline method which other distributions can share, if that is all
they need as well. The secret sauce is put in playbooks/secret.yml,
however the registration code is kept on the kdevops configuration,
this is in keeping with our cloud configuration which lets us keep
credentials on kconfig. This means as-is kconfig files are
confidential. We may change this in the future.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>