vagrant/Kconfig: disable installing libvirt and configuring it
Instead of installing libvirt and configuring it always by default,
add a new option to ask the user if this is the first time they are
running vagrant. Only if so is the option to install libvirt, configure
it and verify it useful. More often than not, people will eventually
use kdevops *many* times on a system, and the defaults should be to skip
this work, given it takes a long time.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>