requirements: update to kdevops v3.3.3
This has a change to let us rely on python3 statically on
the system used for running kdevops. We rely on python3
statically given some distributions such as Debian have
a policy which specifically says not to use /usr/bin/env
and also don't even have /usr/bin/python symlink anymore.
This is the case now for debian testing as of today.

We rely then instead on python3 since our recommended systems
for kdevops to run on is the latest and greatest Debian testing
or OpenSUSE tumbleweed which are *sure* to have python3.

For more details refer to the Debian Python Policy [0].

[0] https://www.debian.org/doc/packaging-manuals/python-policy/ch-python.html#s-interpreter

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>