kdevops: generalize baseline & development kconfig option
We had added a kconfig option to enable the fstests workflow
to reduce the number of hosts used when testing XFS to skip the
development host. This is a generic desirable on kdevops, and so
just bring this forward as a first class kconfig citizen when
configuring kdevops in general.

This lets us also script out the kdevops node generation file.

We bring forward a fix on the preamble file, which was to skip ansible
introduced via commit c090267e50cb88e ("kdevops_nodes.yaml.in: fix
processing ansible enablement when using vagrant"). We forgot to
also fix this on the split files for XFS.

Now with a shared preample/postfix file we can share fixes.

The motivation behind this is twofold:

  o share mode code between node file generation
  o have the ability to easily extend lines for each host entry,
    as we'll need this ability to add some new features

Fixes: c090267e50cb88e ("kdevops_nodes.yaml.in: fix processing ansible enablement when using vagrant")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>