kdevops: add support for annotating lacking support for NVME drives
Since we're a kernel development project, we strive to use the latest
humanely accessible drive technology when possible for for storage,
and when possible we also strive to use the respective virtual solutions
for it.

We currently shoot for using nvme drives for our storage solutions where
we place the extra git trees, development data, etc. Certain
distributions however do not have support for nvme. We need to annotate
this somehow so an alternative can be provided for those older
distributions.

This will enable support for SLES11-SP4 and older, for instance.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>