Initial import: make shared project
Both vagrant and terraform have a need to let you update your
ssh configuration. We use a python script for this, and we used
to fork it. This is not needed, we can just let them use a git
tree as the source, they can use a git subtree. They will use
this git tree as a subtree to get the latest update to the python
script. This way updates are done in only one place.

This is an import of the python script from the terraform module
as it has the latest updates of the Python script. The import is
from the terraform module, terraform-kdevops-add-host-ssh-config [0]
v1.2.0.

$ sha1sum update_ssh_config.py
e1818fec8aee18eb9723869db319b952ec02fef1  update_ssh_config.py

We'll be extending this after the projects are updated to use
this git tree as a subtree.

[0] https://github.com/mcgrof/terraform-kdevops-add-host-ssh-config

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>