Add initial support for reading vagrant ssh-config output
Today the vagrant ansible role relies on calling the python
script once, once to remove hosts only. It then uses the
vagrant ssh-config command to augment your ssh configuration.
This works well for now, however, if we want to *augment* the
information provided by vagrant ssh-config, this won't work
well. For instance, it may be desirable to modify the key
preferences for certain old hosts, otherwise you won't be
able to ssh in to the old hosts. This information is not
provided by vagrant, so we'll later have to add support for
this preference on certain hosts.

We add support to augment the information by also allowing
us to remove and update the ssh configuration for vagrant
users in one command alone. This should allow the ansible
vagrant task to update your ssh configuration to only run
one single command.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>